refactor: move messages from contents to whispering-web
