+++ apps/extension/background/messages/whispering-web/cancelRecording.ts
