+++ apps/extension/background/messages/whispering-web/getSettings.ts
