+++ apps/extension/background/messages/whispering-web/setSettings.ts
