+++ apps/extension/background/messages/whispering-web/toggleRecording.ts
